% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb-get-adm.R
\encoding{UTF-8}
\name{gb_get_adm}
\alias{gb_get_adm}
\alias{gb_get_adm0}
\alias{gb_get_adm1}
\alias{gb_get_adm2}
\alias{gb_get_adm3}
\alias{gb_get_adm4}
\alias{gb_get_adm5}
\title{Retrieve geoBoundaries country files for a specified administrative level}
\source{
geoBoundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
gb_get_adm0(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)

gb_get_adm1(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)

gb_get_adm2(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)

gb_get_adm3(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)

gb_get_adm4(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)

gb_get_adm5(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)
}
\arguments{
\item{country}{Character vector of country identifiers. This can be \code{"all"}
(to return data for all countries), a vector of country names, or ISO3
country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{simplified}{logical. If \code{TRUE}, return simplified boundaries. If
\code{FALSE} (the default), use the premier geoBoundaries release.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"}, or
\code{"gbAuthoritative"}. For most users, \code{"gbOpen"} (the default) is
recommended, as it is CC-BY 4.0 compliant and suitable for most purposes
provided proper attribution is given:
\itemize{
\item \code{"gbHumanitarian"} files are mirrored from
\href{https://www.unocha.org/}{UN OCHA} and may have more restrictive
licensing.
\item \code{"gbAuthoritative"} files are mirrored from UN SALB and cannot be used
for commercial purposes, but are verified through in-country processes.
}}

\item{quiet}{logical. If \code{TRUE}, suppress informational messages.}

\item{overwrite}{logical. If \code{TRUE}, force a fresh download of the source
\code{.zip} file, even if it is already cached.}

\item{path}{Character. Path to a cache directory. If \code{NULL} (the default),
data are stored in the default cache directory (see \code{\link[=gb_set_cache]{gb_set_cache()}}). If
no cache directory has been configured, files are stored in a temporary
directory (see \code{\link[base:tempfile]{base::tempdir()}}).}
}
\value{
An \code{\link[sf:sf]{sf}} object containing the requested boundaries.
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

These functions are convenience wrappers around \code{\link[=gb_get]{gb_get()}} for retrieving
boundaries at different administrative levels:
\itemize{
\item \code{gb_get_adm0()} returns national boundaries.
\item \code{gb_get_adm1()} returns first‑level administrative boundaries
(e.g., states in the United States).
\item \code{gb_get_adm2()} returns second‑level administrative boundaries
(e.g., counties in the United States).
\item \code{gb_get_adm3()} returns third‑level administrative boundaries
(e.g., towns or cities in some countries).
\item \code{gb_get_adm4()} returns fourth‑level administrative boundaries.
\item \code{gb_get_adm5()} returns fifth‑level administrative boundaries.
}

Note that administrative hierarchies vary by country. Use
\link{gb_get_max_adm_lvl} to check the maximum available level.
}
\details{
Individual boundary files in the geoBoundaries database are governed by the
license or licenses specified in their associated metadata (see
\code{\link[=gb_get_metadata]{gb_get_metadata()}}). Users working with individual boundary files should
ensure that they comply with these licenses and cite the original data
sources listed in the metadata. See \strong{Examples}.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

\donttest{
lev2 <- gb_get_adm2(
  c("Italia", "Suiza", "Austria"),
  simplified = TRUE
)


library(ggplot2)

ggplot(lev2) +
  geom_sf(aes(fill = shapeGroup)) +
  labs(
    title = "Second-level administration",
    subtitle = "Selected countries",
    caption = "Source: www.geoboundaries.org"
  )
}
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLoS ONE} 15(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
\code{\link[=gb_get_max_adm_lvl]{gb_get_max_adm_lvl()}}.

Other API functions: 
\code{\link{gb_get}()},
\code{\link{gb_get_world}()}
}
\concept{API functions}
