% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb-get-max-lvl.R
\encoding{UTF-8}
\name{gb_get_max_adm_lvl}
\alias{gb_get_max_adm_lvl}
\title{Retrieve the highest available administrative level for a country}
\source{
geoBoundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
gb_get_max_adm_lvl(
  country = "all",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative")
)
}
\arguments{
\item{country}{Character vector of country identifiers. This can be \code{"all"}
(to return data for all countries), a vector of country names, or ISO3
country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"}, or
\code{"gbAuthoritative"}. For most users, \code{"gbOpen"} (the default) is
recommended, as it is CC-BY 4.0 compliant and suitable for most purposes
provided proper attribution is given:
\itemize{
\item \code{"gbHumanitarian"} files are mirrored from
\href{https://www.unocha.org/}{UN OCHA} and may have more restrictive
licensing.
\item \code{"gbAuthoritative"} files are mirrored from UN SALB and cannot be used
for commercial purposes, but are verified through in-country processes.
}}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing country names and their corresponding
highest available administrative level.
}
\description{
Generate a summary of selected countries—or all countries—and the highest
administrative level available in the geoBoundaries dataset.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

all <- gb_get_max_adm_lvl()
library(dplyr)

# Countries with only 1 level available
all |>
  filter(maxBoundaryType == 1)

# Countries with level 4 available
all |>
  filter(maxBoundaryType == 4)
\dontshow{\}) # examplesIf}
}
\seealso{
Other metadata functions: 
\code{\link{gb_get_metadata}()}
}
\concept{metadata functions}
