% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb-get-world.R
\encoding{UTF-8}
\name{gb_get_world}
\alias{gb_get_world}
\title{Retrieve global composite data (CGAZ) from geoBoundaries}
\source{
geoBoundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
gb_get_world(
  country = "all",
  adm_lvl = "adm0",
  quiet = TRUE,
  overwrite = FALSE,
  path = NULL
)
}
\arguments{
\item{country}{Character vector of country identifiers. This can be \code{"all"}
(to return data for all countries), a vector of country names, or ISO3
country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{adm_lvl}{Type of boundary. Accepted values are administrative
levels 0, 1, and 2. \code{"adm0"} corresponds to national boundaries,
\code{"adm1"} to first‑level subnational boundaries, and \code{"adm2"} to
second‑level boundaries. Uppercase variants (e.g., \code{"ADM1"}) and
numeric values (\code{0}, \code{1}, \code{2}) are also accepted.}

\item{quiet}{logical. If \code{TRUE}, suppress informational messages.}

\item{overwrite}{logical. If \code{TRUE}, force a fresh download of the source
\code{.zip} file, even if it is already cached.}

\item{path}{Character. Path to a cache directory. If \code{NULL} (the default),
data are stored in the default cache directory (see \code{\link[=gb_set_cache]{gb_set_cache()}}). If
no cache directory has been configured, files are stored in a temporary
directory (see \code{\link[base:tempfile]{base::tempdir()}}).}
}
\value{
An \code{\link[sf:sf]{sf}} object containing the requested boundaries.
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

This function returns a global composite for the requested administrative
level, clipped to international boundaries, with gaps filled between
adjacent borders.
}
\details{
Comprehensive Global Administrative Zones (CGAZ) is a set of global
composite administrative boundary products. There are two key differences
between these global composites and individual country downloads:
\itemize{
\item Boundaries undergo extensive simplification to ensure file sizes remain
manageable for typical desktop software.
\item Disputed areas are removed and replaced with polygons defined according
to U.S. Department of State guidelines.
}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

# This download may take some time
\dontrun{
world <- gb_get_world()

library(ggplot2)

ggplot(world) +
  geom_sf() +
  coord_sf(expand = FALSE) +
  labs(caption = "Source: www.geoboundaries.org")
}
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLoS ONE} 15(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{gb_get}()},
\code{\link{gb_get_adm}}
}
\concept{API functions}
