% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gb_max_adm_lvl}
\alias{gb_max_adm_lvl}
\title{Get the highest administrative level available for a given country}
\usage{
gb_max_adm_lvl(country = NULL, type = NULL, version = NULL, license = NULL)
}
\arguments{
\item{country}{characher; a vector of country names or iso3 country code.}

\item{type}{character; defaults to HPSCU. One of HPSCU, HPSCGS, SSCGS, or SSCU.
Determines the type of boundary link you receive. More on details}

\item{version}{character; defaults to the most recent version of geoBoundaries available.}

\item{adm_lvl}{characher; administrative level, adm0, adm1, adm2, adm3, adm4, adm5 or all. adm0 being the country and all to access all available levels.}
}
\value{
a data.frame with the country names and corresponding highest administrative level
}
\description{
Get the highest administrative level available for a given country
}
