% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoboundaries.R, R/utils.R
\name{gb_metadata}
\alias{gb_metadata}
\alias{gb_meta}
\title{Get metadata for a country, administrative level and version}
\usage{
gb_metadata(country, adm_lvl = "adm0", version = "3_0_0")

gb_meta(country = NULL, adm_lvl, type = NULL, version = NULL)
}
\arguments{
\item{country}{characher; a vector of country names}

\item{adm_lvl}{characher; administrative level, adm0, adm1, adm2, adm3, adm4 or adm5. adm0 being the country.}

\item{version}{character; defaults to the most recent version of geoBoundaries available.
The geoboundaries version requested, with underscores.
For example, 3_0_0 would return data from version 3.0.0 of geoBoundaries.}

\item{type}{character; defaults to HPSCU. One of HPSCU, HPSCGS, SSCGS, or SSCU.
Determines the type of boundary link you receive. More on details}
}
\value{
a \code{data.frame} with metadata
}
\description{
Get metadata for a country, administrative level and version

Get metadata for a country and an administrative level
}
