% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoboundaries.R
\name{geoboundaries}
\alias{geoboundaries}
\title{Get the administrative boundaries of selected countries}
\usage{
geoboundaries(country_iso3, adm_level, quiet_download = FALSE)
}
\arguments{
\item{country_iso3}{vector of country iso3}

\item{adm_level}{a character with desired adminadm0, adm1, adm2 or adm3}

\item{quiet_download}{a logical value, if TRUE download are quiet. Default to FALSE}
}
\value{
a sf object
}
\description{
Get the administrative boundaries of selected countries
}
