% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-cache.R
\name{init_rgeoboundaries_cache}
\alias{init_rgeoboundaries_cache}
\title{Initialize 'rgeoboundaries' Cache}
\usage{
init_rgeoboundaries_cache(force = FALSE)
}
\arguments{
\item{force}{Logical. If \code{TRUE}, remove existing cache directory.}
}
\value{
The path to the 'rgeoboundaries' cache directory.
}
\description{
Initialize the 'rgeoboundaries' cache directory.
This function creates the cache directory if it does not exist.

If the \code{force} argument is set to \code{TRUE}, it will remove any existing
cache directory before creating a new one.
}
\examples{
rgeoboundaries_dummy_setup()
init_rgeoboundaries_cache()

init_rgeoboundaries_cache(force = TRUE)
}
\seealso{
Other rgeoboundaries cache management: 
\code{\link{rgeoboundaries_cache_dir}()},
\code{\link{rgeoboundaries_cache_exists}()},
\code{\link{rgeoboundaries_dummy_setup}()},
\code{\link{rm_rgeoboundaries_cache}()}
}
\concept{rgeoboundaries cache management}
