% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-package.R
\docType{package}
\name{rgeoboundaries-package}
\alias{rgeoboundaries-package}
\alias{rgeoboundaries}
\title{rgeoboundaries}
\description{
R client to access geoBoundaries API
}
\author{
\email{team@geoboundaries.org}
}
\keyword{package}
