% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-cache.R
\name{rgeoboundaries_cache_dir}
\alias{rgeoboundaries_cache_dir}
\title{Get 'rgeoboundaries' Cache Directory}
\usage{
rgeoboundaries_cache_dir()
}
\description{
Get the path to the 'rgeoboundaries' cache directory.
This directory is used to store cached data from the 'rgeoboundaries' package.
}
\details{
This directory is managed by R function \code{\link[=R_user_dir]{R_user_dir()}} but can also
be defined in a non-user location by setting ENV variable \code{RGEOBOUNDARIES_CACHE_DIR}
or by setting R option \code{RGEOBOUNDARIES_CACHE_DIR}.
}
\examples{
rgeoboundaries_cache_dir()

options(RGEOBOUNDARIES_CACHE_DIR = tempdir())
rgeoboundaries_cache_dir()
options(RGEOBOUNDARIES_CACHE_DIR = NULL)

Sys.setenv(RGEOBOUNDARIES_CACHE_DIR = tempdir())
rgeoboundaries_cache_dir()
Sys.setenv(RGEOBOUNDARIES_CACHE_DIR = "")
}
\seealso{
Other rgeoboundaries cache management: 
\code{\link{init_rgeoboundaries_cache}()},
\code{\link{rgeoboundaries_cache_exists}()},
\code{\link{rgeoboundaries_dummy_setup}()},
\code{\link{rm_rgeoboundaries_cache}()}
}
\concept{rgeoboundaries cache management}
