% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-cache.R
\name{rgeoboundaries_cache_exists}
\alias{rgeoboundaries_cache_exists}
\title{Check if 'rgeoboundaries' Cache Exists}
\usage{
rgeoboundaries_cache_exists()
}
\description{
Check if the 'rgeoboundaries' cache directory exists.
}
\examples{
rm_rgeoboundaries_cache()
rgeoboundaries_cache_exists()

rgeoboundaries_dummy_setup()
init_rgeoboundaries_cache()
rgeoboundaries_cache_exists()
}
\seealso{
Other rgeoboundaries cache management: 
\code{\link{init_rgeoboundaries_cache}()},
\code{\link{rgeoboundaries_cache_dir}()},
\code{\link{rgeoboundaries_dummy_setup}()},
\code{\link{rm_rgeoboundaries_cache}()}
}
\concept{rgeoboundaries cache management}
