% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-cache.R
\name{rgeoboundaries_dummy_setup}
\alias{rgeoboundaries_dummy_setup}
\title{dummy 'rgeoboundaries' cache}
\usage{
rgeoboundaries_dummy_setup()
}
\description{
dummy 'rgeoboundaries' cache
used for examples or docker environments.
}
\examples{
rgeoboundaries_dummy_setup()
rgeoboundaries_cache_dir()
rm_rgeoboundaries_cache()
}
\seealso{
Other rgeoboundaries cache management: 
\code{\link{init_rgeoboundaries_cache}()},
\code{\link{rgeoboundaries_cache_dir}()},
\code{\link{rgeoboundaries_cache_exists}()},
\code{\link{rm_rgeoboundaries_cache}()}
}
\concept{rgeoboundaries cache management}
