% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-cache.R
\name{rm_rgeoboundaries_cache}
\alias{rm_rgeoboundaries_cache}
\title{Remove 'rgeoboundaries' Cache}
\usage{
rm_rgeoboundaries_cache()
}
\description{
Remove the 'rgeoboundaries' cache directory.
This function deletes the cache directory and all its contents.
}
\examples{
rgeoboundaries_dummy_setup()
rm_rgeoboundaries_cache()
}
\seealso{
Other rgeoboundaries cache management: 
\code{\link{init_rgeoboundaries_cache}()},
\code{\link{rgeoboundaries_cache_dir}()},
\code{\link{rgeoboundaries_cache_exists}()},
\code{\link{rgeoboundaries_dummy_setup}()}
}
\concept{rgeoboundaries cache management}
