% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{gb_metadata}
\title{Deprecated and renamed functions in \pkg{rgeoboundaries} 2.0.0}
\usage{
gb_metadata(
  country = NULL,
  adm_lvl = "all",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative")
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"all"} (that would return the data for all countries), a vector of country
names or ISO3 country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{adm_lvl}{Type of boundary Accepted values are \code{"all"} (all
available boundaries) or the ADM level (\code{"adm0"} is the country boundary,
\code{"adm1"} is the first level of sub national boundaries, \code{"adm2"} is the
second level and so on. Upper case version (\code{"ADM1"}) and the number of
the level (\verb{1,2,3}) and also accepted.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
\code{"gbAuthoritative"}. For most users, we suggest using \code{"gbOpen"}
(the default), as it is CC-BY 4.0 compliant and can be used for most
purposes so long as attribution is provided:
\itemize{
\item \code{"gbHumanitarian"} files are mirrored from
\href{https://www.unocha.org/}{UN OCHA}, but may have less open licensure.
\item \code{"gbAuthoritative"} files are mirrored from
\href{https://salb.un.org/en}{UN SALB}, and cannot  be used for commercial
purposes, but are verified through in-country processes.
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\pkg{rgeoboundaries} 2.0.0 renamed a number of functions to ensure that
every function has a \code{object_verb()} naming scheme and avoiding magical
defaults.
\itemize{
\item \code{gb_metadata()} -> \code{gb_get_metadata()}
}
}
\references{
rOpenSci, Anderson B, Chamberlain S, DeCicco L, Gustavsen J, Krystalli A,
Lepore M, Mullen L, Ram K, Ross N, Salmon M, Vidoni M, Riederer E, Sparks A,
Hollister J (2024).
\emph{rOpenSci Packages: Development, Maintenance, and Peer Review}.
\url{doi:10.5281/zenodo.10797633}.

The tidyverse team. \emph{The Tidyverse Style Guide},
\url{https://style.tidyverse.org/}. Accessed 14 Nov. 2025.

Wickham, Hadley. “Tidy Design Principles.” \emph{Tidyverse.org}, 2025,
\url{https://design.tidyverse.org/}. Accessed 14 Nov. 2025.
}
\concept{deprecated}
\keyword{internal}
