% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{geoboundaries}
\alias{gb_adm0}
\alias{gb_adm1}
\alias{gb_adm2}
\alias{gb_adm3}
\alias{gb_adm4}
\alias{gb_adm5}
\alias{gb_metadata}
\alias{gb_max_adm_lvl}
\title{Deprecated and renamed functions in \CRANpkg{rgeoboundaries} 2.0.0}
\usage{
geoboundaries(
  country = NULL,
  adm_lvl = "adm0",
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_adm0(
  country = NULL,
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_adm1(
  country = NULL,
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_adm2(
  country = NULL,
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_adm3(
  country = NULL,
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_adm4(
  country = NULL,
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_adm5(
  country = NULL,
  type = "unsimplified",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  quiet = TRUE,
  overwrite = FALSE,
  version = deprecated()
)

gb_metadata(
  country = NULL,
  adm_lvl = "all",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative")
)

gb_max_adm_lvl(
  country = NULL,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative")
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"all"} (that would return the data for all countries), a vector of country
names or ISO3 country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{adm_lvl}{Type of boundary Accepted values are \code{"all"} (all
available boundaries) or the ADM level (\code{"adm0"} is the country boundary,
\code{"adm1"} is the first level of sub national boundaries, \code{"adm2"} is the
second level and so on. Upper case version (\code{"ADM1"}) and the number of
the level (\verb{1, 2, 3, 4, 5}) and also accepted.}

\item{type}{character. One of \code{"simplified"} and \code{"unsimplified"}. Other
values would be assumed as \code{"unsimplified"}.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
\code{"gbAuthoritative"}. For most users, we suggest using \code{"gbOpen"}
(the default), as it is CC-BY 4.0 compliant and can be used for most
purposes so long as attribution is provided:
\itemize{
\item \code{"gbHumanitarian"} files are mirrored from
\href{https://www.unocha.org/}{UN OCHA}, but may have less open licensure.
\item \code{"gbAuthoritative"} files are mirrored from
\href{https://salb.un.org/en}{UN SALB}, and cannot  be used for commercial
purposes, but are verified through in-country processes.
}}

\item{quiet}{logical. If \code{TRUE} suppresses informational messages.}

\item{overwrite}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.zip} file.}

\item{version}{Deprecated argument.}
}
\value{
These functions are re-directed to their replacement, providing the same
output.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\CRANpkg{rgeoboundaries} 2.0.0 renamed a number of functions to ensure that
every function has a \code{object_verb()} naming scheme and avoiding magical
defaults.
\itemize{
\item \code{geoboundaries()} -> \code{\link[=gb_get]{gb_get()}} and \code{\link[=gb_get_world]{gb_get_world()}}.
\item \verb{gb_adm*} family -> \code{\link[=gb_get_adm]{gb_get_adm*}} family.
\item \code{gb_metadata()} -> \code{\link[=gb_get_metadata]{gb_get_metadata()}}.
\item \code{gb_max_adm_lvl()} -> \code{\link[=gb_get_max_adm_lvl]{gb_get_max_adm_lvl()}}.
}

These functions are likely to be removed in the future.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
# Show deprecation messages
gb_metadata()
\dontshow{\}) # examplesIf}
}
\references{
rOpenSci, Anderson B, Chamberlain S, DeCicco L, Gustavsen J, Krystalli A,
Lepore M, Mullen L, Ram K, Ross N, Salmon M, Vidoni M, Riederer E, Sparks A,
Hollister J (2024).
\emph{rOpenSci Packages: Development, Maintenance, and Peer Review}.
\url{doi:10.5281/zenodo.10797633}.

The tidyverse team. \emph{The Tidyverse Style Guide},
\url{https://style.tidyverse.org/}. Accessed 14 Nov. 2025.

Wickham, Hadley. “Tidy Design Principles.” \emph{Tidyverse.org}, 2025,
\url{https://design.tidyverse.org/}. Accessed 14 Nov. 2025.
}
\concept{deprecated}
\keyword{internal}
