% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{gb_cache}
\alias{gb_cache}
\alias{gb_set_cache}
\alias{gb_get_cache}
\alias{gb_clear_cache}
\alias{gb_list_cache}
\alias{gb_delete_from_cache}
\title{Caching geoBoundaries downloaded files}
\usage{
gb_set_cache(path)

gb_get_cache(create = FALSE)

gb_clear_cache(force = FALSE)

gb_list_cache(full_path = FALSE)

gb_delete_from_cache(file)
}
\arguments{
\item{path}{character; path of the directory to set}

\item{create}{logical; if TRUE create missing cache}

\item{force}{logical; force delete. default: \code{FALSE}}

\item{file}{Character, the file to delete}
}
\value{
the cache directory

list of files in the cache
}
\description{
Manage cached geoBoundaries downloaded files
}
\details{
The default cache directory is
\verb{~/.cache/R/gb_cache}, but you can set
your own path using \code{gb_set_cache(path)}
}
\note{
This function will clear all cached files
}
\examples{
\dontrun{
gb_cache
## change the default cache directory
tmp <- tempdir()
gb_set_cache(tmp)

## print current cache directory
gb_get_cache()

## List available files in the current cache directory
gb_list_cache()

l <- gb_list_cache()[1] ## get the first file
gb_delete_from_cache(l) ## delete it

gb_clear_cache() ## delete all cached files
}
}
