% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gb_max_adm_lvl}
\alias{gb_max_adm_lvl}
\title{Get the highest administrative level available for a given country}
\usage{
gb_max_adm_lvl(country = NULL, type = NULL)
}
\arguments{
\item{country}{characher; a vector of country names or iso3 country code.}

\item{type}{character; defaults to simplified. One of 'simplified' or 'unsimplified'.
Determines the type of boundary link you receive. More on details}
}
\value{
a data.frame with the country names and corresponding highest administrative level
}
\description{
Get the highest administrative level available for a given country
}
