% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gb_metadata}
\alias{gb_metadata}
\title{Get metadata for a country, administrative level, type of data and version}
\usage{
gb_metadata(country = NULL, adm_lvl = "all", type = NULL)
}
\arguments{
\item{country}{characher; a vector of country names or iso3 country code.}

\item{adm_lvl}{characher; administrative level, adm0, adm1, adm2, adm3, adm4, adm5 or all. adm0 being the country and all to access all available levels.}

\item{type}{character; defaults to simplified. One of 'simplified' or 'unsimplified'.
Determines the type of boundary link you receive. More on details}
}
\description{
Get metadata for a country and an administrative level, type of data and version
}
