% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gb_metadata}
\alias{gb_metadata}
\title{Get metadata for a country and an administrative level}
\usage{
gb_metadata(country, adm_level)
}
\arguments{
\item{country}{characher; a vector of country names}

\item{adm_level}{characher; administrative level, adm0, adm1, adm2, adm3, adm4 or adm5. admO being the country.}
}
\description{
Get metadata for a country and an administrative level
}
