% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoboundaries.R
\name{geoboundaries}
\alias{geoboundaries}
\alias{gb_adm0}
\alias{gb_adm1}
\alias{gb_adm2}
\alias{gb_adm3}
\alias{gb_adm4}
\alias{gb_adm5}
\title{Get the administrative boundaries of selected countries}
\usage{
geoboundaries(country, adm_lvl = "adm0", quiet = TRUE)

gb_adm0(country, quiet = TRUE)

gb_adm1(country, quiet = TRUE)

gb_adm2(country, quiet = TRUE)

gb_adm3(country, quiet = TRUE)

gb_adm4(country, quiet = TRUE)

gb_adm5(country, quiet = TRUE)
}
\arguments{
\item{country}{characher; a vector of country names}

\item{adm_lvl}{characher; administrative level, adm0, adm1, adm2, adm3, adm4 or adm5. admO being the country.}

\item{quiet}{logical; if TRUE download are quiet. Default to FALSE}
}
\value{
a \code{sf} object
}
\description{
Access country boundaries at a specified administrative level
}
\details{
\itemize{
\item \code{gb_adm0} returns the country boundaries
\item \code{gb_adm1} if available, returns the country first administrative level boundaries
\item \code{gb_adm2} if available, returns the country second administrative level boundaries
\item \code{gb_adm3} if available, returns the country third administrative level boundaries
\item \code{gb_adm4} if available, returns the country fourth administrative level boundaries
\item \code{gb_adm5} if available, returns the country first administrative level boundaries
}
}
