% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-package.R
\docType{package}
\name{rgeoboundaries-package}
\alias{rgeoboundaries-package}
\alias{rgeoboundaries}
\title{rgeoboundaries}
\description{
R client to access geoBoundaries API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmgeolab/rgeoboundaries}
  \item Report bugs at \url{https://github.com/wmgeolab/rgeoboundaries/issues}
}

}
\author{
\email{mail@ahmadoudicko.com}
}
\keyword{package}
