% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoboundaries-package.R
\docType{package}
\name{rgeoboundaries-package}
\alias{rgeoboundaries}
\alias{rgeoboundaries-package}
\title{rgeoboundaries: A Client to geoBoundaries, a Political Administrative Boundaries Dataset}
\description{
Client the access data from the geoBoundaries API who provides country political administrative boundaries dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmgeolab/rgeoboundaries}
  \item Report bugs at \url{https://github.com/wmgeolab/rgeoboundaries/issues}
}

}
\author{
\strong{Maintainer}: Daniel Runfola \email{team@geoboundaries.org} [contributor]

Authors:
\itemize{
  \item Ahmadou Dicko \email{mail@ahmadoudicko.com}
}

Other contributors:
\itemize{
  \item Andy South \email{southandy@gmail.com} [contributor]
  \item Rohith Reddy Mandala \email{rohithreddy.m4@gmail.com} [contributor]
}

}
\keyword{internal}
